Agregando un poco de estilo

+++ Sistema_automóviles/Forms/Frm_ventas.cs
@@ -10,6 +10,7 @@ using System.Threading.Tasks;
 using System.Windows.Controls;
 using System.Windows.Forms;
 using Sistema_automóviles.Models.Classes;
+using System.Windows.Controls.Primitives;
 
 namespace Sistema_automóviles.Forms
 {
@@ -19,6 +20,7 @@ namespace Sistema_automóviles.Forms
         public Frm_ventas()
         {
             InitializeComponent();
+            StartPosition = FormStartPosition.CenterParent;
         }
 
         private void Frm_ventas_Load(object sender, EventArgs e)
@@ -26,26 +28,25 @@ namespace Sistema_automóviles.Forms
             DGV.DataSource = db.GetAllVentas();
 
             DataGridViewTextBoxColumn nuevaColumna = new DataGridViewTextBoxColumn();
-            nuevaColumna.HeaderText = "Eliminar";
+            nuevaColumna.HeaderText = "X";
             nuevaColumna.Name = "Eliminar";
+            nuevaColumna.AutoSizeMode = DataGridViewAutoSizeColumnMode.None;
+            nuevaColumna.Width = 40;
             DGV.Columns.Add(nuevaColumna);
-            foreach (DataGridViewRow row in DGV.Rows)
-            {
-                DataGridViewCell cell = row.Cells["Eliminar"];
-                if (cell != null)
-                {
-                    cell.Style.BackColor = Color.FromArgb(255, 204, 204);
-                }
-            }
+            //foreach (DataGridViewRow row in DGV.Rows)
+            //{
+            //    DataGridViewCell cell = row.Cells["Eliminar"];
+            //    cell.Style.BackColor = Color.FromArgb(255, 255, 0, 0);
+            //}
         }
 
         private void DGV_CellClick(object sender, DataGridViewCellEventArgs e)
         {
             int id = Convert.ToInt32(DGV.CurrentRow.Cells[0].Value);
-            if (e.ColumnIndex==7)
+            if (e.ColumnIndex == 7 && e.RowIndex >= 0)
             {
                 DialogResult result = MessageBox.Show("¿Desea eliminar la venta", "Confirmación", MessageBoxButtons.OKCancel, MessageBoxIcon.Warning);
-                if(result == DialogResult.OK)
+                if (result == DialogResult.OK)
                 {
                     if (db.DeleteVenta(id))
                         DGV.DataSource = db.GetAllVentas();
@@ -53,13 +54,16 @@ namespace Sistema_automóviles.Forms
             }
             else
             {
-                (Proveedor, Cliente) datos = db.GetInfoVenta(id);
-                var p = datos.Item1;
-                var c = datos.Item2;
-                MessageBox.Show($"\tCLIENTE\n-Nombre: {c.Nombre} {c.APaterno} {c.AMaterno}\n-Correo: {c.Correo}\n" +
-                    $"-Teléfono: {c.Telefono}\n-RFC: {c.RFC}\n-CP: {c.CP}\n-------------------------------------------" +
-                    $"\n\tPROVEEDOR\n-Nombre: {p.Nombre}\n-Teléfono: {p.Telefono}\n-Correo: {p.Correo}" +
-                    $"\n-Dirección: {p.Direccion}", "Información de la venta", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                if (e.RowIndex >= 0)
+                {
+                    (Proveedor, Cliente) datos = db.GetInfoVenta(id);
+                    var p = datos.Item1;
+                    var c = datos.Item2;
+                    MessageBox.Show($"\tCLIENTE\n-Nombre: {c.Nombre} {c.APaterno} {c.AMaterno}\n-Correo: {c.Correo}\n" +
+                        $"-Teléfono: {c.Telefono}\n-RFC: {c.RFC}\n-CP: {c.CP}\n-------------------------------------------" +
+                        $"\n\tPROVEEDOR\n-Nombre: {p.Nombre}\n-Teléfono: {p.Telefono}\n-Correo: {p.Correo}" +
+                        $"\n-Dirección: {p.Direccion}", "Información de la venta", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                }
             }
         }
     }
